Optimizing code and ading a 'screen' version
Code optimization and organizing comments in the master file :  
fbx-delta-nba_bash_api.sh

Cloning master file to fbx-delta-nba_bash_api_screen.sh and adding screen support to launch websocket connection inside a screen (as a workaround):

Websocket API is used to retrieve Virtual Machine console and for text editor (like vim) to work in the VM through the websocket API, I have to use 'stty'. But the way I have to use it block the trap of SIGINT ==> If you want to close the websocket connection after logout the VM console, you need to kill it from another terminal 
That why I add screen support in this second file. 

Now, the websocket connection is launched inside a screen and to close the connection, you simply need to close the screen (CTRL-A+K) .

Enjoy !
nbanba